chore(artifacts): move GCS reference tests to the unit suite
The GCS reference tests build a local Artifact and swap in a stub GCS
client. They never use the wandb backend, so they do not need to run in
the system-test job.

Move these tests, along with the mock_gcs helper they use, into a new
tests/unit_tests/test_artifacts/test_gcs_references.py. The new module
defines its own small artifact fixture; the system-test module keeps its
copy for the tests that remain there.

Also drop the now-unused ArtifactState and gcs_handler imports from the
system-test module. This is a pure move with no change in test behavior.
